+++ app/src/main/java/com/damsoft/overheidsdata/adapter/ThemesAdapter.kt
@@ -1,23 +1,24 @@
-package com.damsoft.overheidsdata.adapter.viewholder
+package com.damsoft.overheidsdata.adapter
 
 import android.support.v7.widget.RecyclerView
 import android.view.ViewGroup
 import com.damsoft.overheidsdata.R
+import com.damsoft.overheidsdata.adapter.viewholder.ThemeViewHolder
 import com.damsoft.overheidsdata.db.ThemeEntity
 import com.damsoft.overheidsdata.inflate
 
 /**
  * Created by abhinav.sharma on 01/11/17.
  */
-class ThemesAdapter(private val listener: (ThemeEntity) -> Unit, private var sourceList: List<ThemeEntity>) : RecyclerView.Adapter<ThemeViewHolder>() {
+class ThemesAdapter(private val listener: (ThemeEntity) -> Unit, private var themesList: List<ThemeEntity>) : RecyclerView.Adapter<ThemeViewHolder>() {
 
 
     override fun getItemCount(): Int {
-        return sourceList.size
+        return themesList.size
     }
 
     private fun getItem(position: Int): ThemeEntity {
-        return sourceList[position]
+        return themesList[position]
     }
 
     override fun onBindViewHolder(holder: ThemeViewHolder, position: Int) =
@@ -27,7 +28,7 @@ class ThemesAdapter(private val listener: (ThemeEntity) -> Unit, private var sou
             ThemeViewHolder(parent.inflate(R.layout.layout_theme_single))
 
     fun updateDataSet(data: List<ThemeEntity>){
-        sourceList = data
+        themesList = data
         notifyDataSetChanged()
     }
 }
